Start big move
This is the first of many commits which mostly move things

Annoyingly the source files need and then don't need annotations to keep tests passing.
It never communicates failures so I just make what seem like reasonable changes until it passes
