+++ lib/server-impl.js
@@ -6,7 +6,7 @@ const migrator = require('../migrator');
 const getApp = require('./app');
 
 const { startMonitoring } = require('./metrics');
-const { createStores } = require('./db');
+const { createStores } = require('./storage');
 const { createOptions } = require('./options');
 const StateService = require('./state-service');
 const User = require('./user');
