+++ lib/server-impl.test.js
@@ -23,7 +23,7 @@ const serverImpl = proxyquire('./server-impl', {
             return o;
         },
     },
-    './db': {
+    './storage': {
         createStores() {
             return {
                 eventStore,
