+++ lib/storage/client-applications-store.js
