+++ lib/storage/client-instance-store.js
