+++ lib/storage/client-metrics-store.test.js
