+++ lib/storage/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { createDb } = require('./db-pool');
+const { createDb } = require('./db/db-pool');
 const EventStore = require('./event-store');
 const FeatureToggleStore = require('./feature-toggle-store');
 const StrategyStore = require('./strategy-store');
