+++ lib/storage/strategy-store.js
