+++ lib/storage/utils/const/application-store.js
@@ -1,4 +1,3 @@
-/* eslint camelcase:off */
 'use strict';
 
 const COLUMNS = [
