+++ lib/storage/utils/const/event-store.js
@@ -1,4 +1,3 @@
-/* eslint camelcase:off */
 'use strict';
 
 const TABLE = 'events';
