+++ lib/storage/utils/const/instance-store.js
@@ -1,4 +1,3 @@
-/* eslint camelcase:off */
 'use strict';
 
 const COLUMNS = [
