+++ lib/storage/utils/const/metrics-store.js
@@ -1,4 +1,3 @@
-/* eslint camelcase:off */
 'use strict';
 
 const COLUMNS = ['id', 'created_at', 'metrics'];
