+++ lib/storage/utils/const/strategy-store.js
