+++ lib/storage/utils/mappings/client-application-store.js
