+++ lib/storage/utils/mappings/client-instance-store.js
